# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Curator v2 API
  version: 1.10.0
  description: |
    C4 of the frozen contracts (design doc 03). JSON in, JSON out, UTF-8.
    Every route lives under the mount prefix `{base}` (empty by default, `/curation` in
    production; the gateway does not strip it). Write operations are asynchronous: they
    return the resource and its state, progress arrives over SSE or polling.
    Errors use one body, `Error`; `code` is stable for machines, `message` is Chinese for people.
    `precheck_failed` (422, any write that starts work) carries `error.details` shaped as
    `PrecheckDetails`: one entry per pre-start check (D30).
    Responses that point people at a page carry `links` (D22): absolute URLs when
    `server.publicBaseUrl` is configured, otherwise relative paths with `absolute: false`.
    Write endpoints accept only `application/json` - send the header on every write, even without
    a body - and cross-site writes are refused; there is no CORS.
    Ids are opaque strings. Records made since 1.10.0 get `<prefix>-<9 lowercase letters>` (task,
    sub, ds, pf, cred, vb, vm, upl; e.g. `task-kqzmrtbwe`, D45); records made before keep their
    `<prefix>_...` id, and both work everywhere. Ids carry no time: lists sort by creation time.

    1.1.0 (2026-09-21): dataset registry (D36): `/datasets` now lists registered datasets and the
    TOS / cache-bucket listing moved to `/datasets/browse`; `/overview`; starting a task re-checks
...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /credentials:
    get:
      tags: [credentials]
      operationId: listCredentials
      summary: List TOS access keys (never returns secrets)
      responses:
        "200":
          description: all access keys of the owner
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/Credential"}}
        default: {$ref: "#/components/responses/Error"}
    post:
      tags: [credentials]
      operationId: createCredential
      summary: Save an access key; verifies identity only (D30, doc 08 §4)
      description: |
        Saving never fails because verification failed: the key is stored and marked
        `failed` or `unverified`. Read and write permissions are checked against concrete
        buckets when a task starts.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CredentialCreate"}
      responses:
        "201":
          description: saved
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Credential"}
        default: {$ref: "#/components/responses/Error"}
  /credentials/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    put:
      tags: [credentials]
      operationId: updateCredential
      summary: Update; a secret left empty stays unchanged
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CredentialUpdate"}
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Credential"}
        default: {$ref: "#/components/responses/Error"}
    delete:
      tags: [credentials]
      operationId: deleteCredential
      summary: Delete (409 credential_in_use while a non-terminal task uses it)
      description: |
        Referenced only by finished tasks: requires `confirm=true`. Those tasks keep their
        reports, which then say the key is gone; rebind-credentials gives them another one.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: confirm
          in: query
          schema: {type: boolean, default: false}
      responses:
        "204": {description: deleted}
        default: {$ref: "#/components/responses/Error"}
  /credentials/{id}/verify:
    parameters:
      - $ref: "#/components/parameters/PathId"
    post:
      tags: [credentials]
      operationId: verifyCredential
      summary: Verify identity again
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: verification result (also stored on the key)
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VerifyResult"}
        default: {$ref: "#/components/responses/Error"}

  # ---------------------------------------------------------------- vlm backends
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic against htpasswd (bcrypt / apr1), realm "Robot Data Curation" (doc 08 §5)
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: {type: string}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: the same key within 24 hours returns the first response (doc 03 §8)
      schema: {type: string, minLength: 8, maxLength: 128}
...
  responses:
    Error:
      description: |
        400 validation_failed, 401 unauthorized, 404 not_found, 409 task_state_conflict /
        subtask_active / credential_in_use / backend_in_use / name_taken / preflight_expired /
        source_changed / result_changed / idempotency_conflict, 412 precondition_failed,
        422 precheck_failed / confirm_path_mismatch / model_check_failed, 500 internal
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              enum: [validation_failed, unauthorized, not_found, task_state_conflict, subtask_active,
                     credential_in_use, backend_in_use, dataset_in_use, name_taken, preflight_expired, precheck_failed,
                     source_changed, result_changed, confirm_path_mismatch, model_check_failed,
                     idempotency_conflict, precondition_failed, method_not_allowed, internal]
            message: {type: string, description: "Chinese, shown to people as is"}
            details: {type: object}
...
    Region:
      type: string
      pattern: "^[a-z0-9][a-z0-9-]{1,31}$"
...
    VerifyState:
      enum: [unverified, ok, failed]
    VerifyResult:
      type: object
      additionalProperties: false
      required: [verify_state, last_verified_at]
      properties:
        verify_state: {$ref: "#/components/schemas/VerifyState"}
        last_verified_at: {type: integer}
        error: {type: [string, "null"]}
...
    Credential:
      type: object
      additionalProperties: false
      required: [id, name, kind, meta, verify_state, created_at, updated_at]
      properties:
        id: {type: string}
        name: {type: string}
        kind: {const: tos}
        meta:
          type: object
          additionalProperties: false
          required: [region]
          properties:
            region: {$ref: "#/components/schemas/Region"}
            endpoint: {type: string}
            test_bucket: {type: string, description: "HeadBucket target when the key cannot ListBuckets"}
            access_key_id_hint: {type: string, description: "last 4 characters only"}
        verify_state: {$ref: "#/components/schemas/VerifyState"}
        last_verified_at: {type: [integer, "null"]}
        last_verify_error: {type: [string, "null"]}
        references:
          type: object
          properties:
            active_tasks: {type: integer, minimum: 0}
            historical_tasks: {type: integer, minimum: 0}
        created_at: {type: integer}
        updated_at: {type: integer}
    CredentialCreate:
      type: object
      additionalProperties: false
      required: [name, access_key_id, secret_access_key, region]
      properties:
        name: {type: string, minLength: 1, maxLength: 64}
        access_key_id: {type: string, minLength: 1}
        secret_access_key: {type: string, minLength: 1, writeOnly: true}
        region: {$ref: "#/components/schemas/Region"}
        endpoint: {type: string}
        test_bucket: {type: string}
        session_token: {type: string, writeOnly: true, description: "temporary credentials only; empty = none"}
    CredentialUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 64}
        access_key_id: {type: string}
        secret_access_key: {type: string, writeOnly: true, description: "empty = unchanged"}
        region: {$ref: "#/components/schemas/Region"}
        endpoint: {type: string}
        test_bucket: {type: string}
        session_token: {type: string, writeOnly: true, description: "temporary credentials only; empty = none"}

    # ---------- vlm
...